supportcloudfunction/requirements.txt
google-cloud-datastore
google-cloud-storage
